requirements.txt
taipy
scikit-learn
numpy
pandas
pillow